feat: change module name

+++ lib/disk/disk_manager_test.go
@@ -25,8 +25,8 @@ func TestReadWriteFile(t *testing.T) {
 	if err != nil {
 		t.Error(err)
 	}
-	assert.Equal(t, 1, pageReader.GetInt(0))
-	assert.Equal(t, 2, pageReader.GetInt(4))
-	assert.Equal(t, 3, pageReader.GetInt(8))
+	assert.Equal(t, int32(1), pageReader.GetInt(0))
+	assert.Equal(t, int32(2), pageReader.GetInt(4))
+	assert.Equal(t, int32(3), pageReader.GetInt(8))
 	assert.Equal(t, "lintang", pageReader.GetString(12))
 }
